// src/enums/genericTypeEnum.ts
export enum GenericTypeEnum {
  None = 0,

  // Start Light generic type
  // Actions
  // LIGHT_TOGGLE
  LightToggle = 1,

  //LIGHT_ON
  LightOn = 2,

  // LIGHT_OFF
  LightOff = 3,

  // LIGHT_SLIDER
  LightSlider = 4,

  // LIGHT_SET_COLOR
  LightSetColor = 5,

  // LIGHT_MODE
  LightMode = 6,

  // LIGHT_SET_COLOR_TEMP
  LightSetColorTemp = 7,

  //Infos
  // LIGHT_STATE
  LightState = 8,

  // LIGHT_STATE_BOOL
  LightStateBool = 9,

  // LIGHT_BRIGHTNESS
  LightBrightness = 10,

  // LIGHT_COLOR
  LightColor = 11,

  // LIGHT_COLOR_TEMP
  LightColorTemp = 12,
  // End Light generic type
}